pyproject.toml
[project]
name = "revobank-api"
version = "0.1.0"
description = "RESTful API for RevoBank application"
readme = "README.md"
authors = [
    {name = "RevoBank Team"}
]
requires-python = ">=3.9,<4.0"
dependencies = [
    "bcrypt>=4.3.0",
    "flask>=3.1.0",
    "flask-cors>=4.0.2",
    "gunicorn>=23.0.0",
    "pydantic[email]>=2.10.6",
    "pyjwt>=2.10.1",
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "python-dotenv>=1.1.0",
    "ruff>=0.9.9",
    "sqlalchemy>=2.0.39",
    "werkzeug>=3.1.3",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.uv]
# UV-specific settings can go here if needed in the future
